# perf: speed up synthetic metric generation